refactor(shopper): CHECKOUT-2951 Convert `customer` module into TypeScript
